pyproject.toml
[project]
name = "company_claims_analysis"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Humphrey Yeboah",email = "haky-works@gmail.com"},
]
dependencies = [
    "pendulum>=3.0.0",
    "pandas>=2.2.3",
    "xlsxwriter>=3.2.0",
    "streamlit>=1.40.2",
    "plotly>=5.24.1",
    "numpy>=2.1.3",
    "seaborn>=0.13.2",
    "openpyxl>=3.1.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.9.3",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "pip>=24.3.1",
    "ipykernel>=6.29.5",
    "pytest>=8.3.3",
    "Faker>=33.1.0",
    "hypothesis>=6.122.0",
    "factory-boy>=3.3.1",
]
